Allow data to be ingested from an S3 bucket rather than local files
Data is now ingested to the OpenSearch instance from an S3 bucket instead of from local files.
The S3 bucket name is specified in the .env file, for now. Currently, the script only indexes
comment files ending with .json to an index named comments, and does so for all dockets listed
in the S3 bucket.